Fixed stack overflow issue using smaller stacks (1MB).
Windows has a maximum stack size of 1MB or so by default and that was
causing the emulator to crash. This was fixed by making Gba::new inline
so that the stack allocations and subsequent moves to the heap could be
optimized into just a single heap allocation. It also now provides
Gba::alloc which will just return a boxed Gba.